chore(storage): restructure storage module exports with index files and interface separation
