+++ src/libs/storage/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './storage.interface';
+export * from './upload-file.types';
