+++ src/libs/storage/interfaces/upload-file.types.ts
@@ -1,0 +1,15 @@
+export interface StorageUploadParams {
+  fileBuffer: Buffer;
+  filename: string;
+  mimetype: string;
+  folder?: string;
+  resourceType?: 'image' | 'video' | 'raw';
+}
+
+export interface StorageFileMetadata {
+  url: string;
+  publicId: string;
+  size?: number;
+  format?: string;
+  resourceType?: 'image' | 'video' | 'raw';
+}
